pyproject.toml
[project]
name = "ororatech-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "glob2 (>=0.7,<0.8)",
    "keras (>=3.10.0,<4.0.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "ororatech_challenge", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
